requirements.txt
flake8==7.2.0
jinja2==3.1.6
pytest==8.3.5
pytest-cov==6.1.1
pyyaml==6.0.2
